Добавлен новый класс для запросов по названию города.
